[#112] :sparkles: refactor: router lib 폴더로 이동
