+++ client/src/index.tsx
@@ -1,15 +1,22 @@
 import App from '@/layouts/App';
 import React from 'react';
 import { render } from 'react-dom';
-import { Router } from '@/core/Router';
+import { Router } from '@/lib/Router';
 import GlobalStyle from '@/styles/globalStyle';
+import { QueryClient, QueryClientProvider } from 'react-query';
+import { ReactQueryDevtools } from 'react-query/devtools';
+
+const queryClient = new QueryClient();
 
 render(
   <>
     <GlobalStyle />
-    <Router>
-      <App />
-    </Router>
+    <QueryClientProvider client={queryClient}>
+      <Router>
+        <App />
+      </Router>
+      <ReactQueryDevtools initialIsOpen />
+    </QueryClientProvider>
   </>,
   document.getElementById('app')
 );
