+++ client/src/pages/Detail/Detail.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import * as S from './styles';
-import { useParams } from '@/core/Router';
+import { useParams } from '@/lib/Router';
 import { ProductDetailTab } from '@/components/Shared/Tab';
 import ProductInfo from '@/components/ProductDetails/ProductInfo';
