+++ client/src/pages/Login/Login.tsx
@@ -2,10 +2,18 @@ import React from 'react';
 import { Input } from '@/components/Shared/Input';
 import Button from '@/components/Shared/Button';
 import * as S from './styles';
-import { Link, useHistory } from '@/core/Router';
+import { Link, useHistory } from '@/lib/Router';
+import { getCurrentUser } from '@/lib/api/user/getCurrentUser';
 import axios from 'axios';
 
+import { useQuery } from 'react-query';
+
 const Login = () => {
+  const { isLoading, error, data, isFetching } = useQuery(
+    'user',
+    getCurrentUser
+  );
+
   const { historyPush } = useHistory();
   const GithubLogin = async () => {
     //const data = await axios.get('api/product');
@@ -13,6 +21,9 @@ const Login = () => {
     const data = await axios.get('api/auth');
     window.location.href = data.data;
   };
+
+  // if (data) return;
+
   return (
     <S.LoginForm>
       <S.LoginTitle level={4}>회원 로그인</S.LoginTitle>
